+++ components/auth/org.wso2.carbon.auth.token.introspection/src/main/java/org/wso2/carbon/auth/token/introspection/IntrospectionManager.java
@@ -15,13 +15,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.carbon.auth.introspection;
+package org.wso2.carbon.auth.token.introspection;
 
-import org.wso2.carbon.auth.introspection.dto.IntrospectionContext;
+import org.wso2.carbon.auth.token.introspection.dto.IntrospectionResponse;
 
 /**
- * Token validation handler interface
+ * Introspection manager interface
  */
-public interface TokenValidatorHandler {
-    void validate(IntrospectionContext context) throws IntrospectionException;
+public interface IntrospectionManager {
+    IntrospectionResponse introspect(String token);
 }
