+++ components/auth/org.wso2.carbon.auth.token.introspection/src/main/java/org/wso2/carbon/auth/token/introspection/dto/IntrospectionContext.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.carbon.auth.introspection.dto;
+package org.wso2.carbon.auth.token.introspection.dto;
 
 /**
  * Introspection context holding class
