+++ features/auth-features/pom.xml
@@ -36,6 +36,9 @@
         <module>org.wso2.carbon.auth.core.feature</module>
         <module>org.wso2.carbon.auth.client.registration.rest.api.feature</module>
         <module>org.wso2.carbon.auth.token.introspection.rest.api.feature</module>
+        <module>org.wso2.carbon.auth.oauth.rest.api.feature</module>
+        <module>org.wso2.carbon.auth.user.mgt.rest.api.feature</module>
+        <module>org.wso2.carbon.auth.user.store.feature</module>
     </modules>
 
 </project>
